requirements.txt
apturl0.5.2             
asgiref3.3.1               
astroid2.4.2               
attrs20.3.0              
Automat20.2.0              
bcrypt3.1.7               
blinker1.4                 
Brlapi0.7.0               
certifi2019.11.28          
chardet3.0.4               
chrome-gnome-shell0.0.0               
Click7.0                 
colorama0.4.3               
command-not-found0.3                 
constantly15.1.0              
cryptography2.8                 
cssselect1.1.0               
cupshelpers1.0                 
dbus-python1.2.16              
defer1.0.6               
distro1.4.0               
distro-info0.23ubuntu1         
Django3.1.6               
duplicity0.8.12.0            
entrypoints0.3                 
fasteners0.14.1              
future0.18.2              
httplib20.14.0              
hyperlink21.0.0              
idna2.8                 
incremental17.5.0              
isort5.7.0               
itemadapter0.2.0               
itemloaders1.0.4               
jmespath0.10.0              
keyring18.0.1              
language-selector0.1                 
launchpadlib1.10.13             
lazr.restfulclient0.14.2              
lazr.uri1.0.3               
lazy-object-proxy1.4.3               
lockfile0.12.2              
louis3.12.0              
lxml4.6.2               
macaroonbakery1.3.1               
Mako1.1.0               
MarkupSafe1.1.0               
mccabe0.6.1               
monotonic1.5                 
netifaces0.10.4              
oauthlib3.1.0               
olefile0.46                
paramiko2.6.0               
parsel1.6.0               
pexpect4.6.0               
Pillow7.0.0               
pip20.0.2              
Protego0.1.16              
protobuf3.6.1               
pyasn10.4.8               
pyasn1-modules0.2.8               
pycairo1.16.2              
pycups1.9.73              
PyDispatcher2.0.5               
PyGObject3.36.0              
PyHamcrest2.0.2               
PyJWT1.7.1               
pylint2.6.0               
pymacaroons0.13.0              
PyNaCl1.3.0               
pyOpenSSL20.0.1              
pyRFC33391.1                 
python-apt2.0.0
python-dateutil2.7.3               
python-debian0.1.36ubuntu1       
pytz2019.3              
pyxdg0.26                
PyYAML5.3.1               
queuelib1.5.0               
reportlab3.5.34              
requests2.22.0              
requests-unixsocket0.2.0               
Scrapy2.4.1               
SecretStorage2.3.1               
selenium3.141.0             
service-identity18.1.0              
setuptools45.2.0              
simplejson3.16.0              
six1.14.0              
sqlparse0.4.1               
ssh-import-id5.10                
systemd-python234                 
toml0.10.2              
Twisted20.3.0              
ubuntu-advantage-tools20.3                
ubuntu-drivers-common0.0.0               
ufw0.36                
unattended-upgrades0.1                 
urllib31.25.8              
usb-creator0.3.7               
w3lib1.22.0              
wadllib1.3.3               
wheel0.34.2              
wrapt1.12.1              
xkit0.0.0               
zope.interface5.2.0   
ubuntu0.20.4.4  
